Refactor autoprecompiles (#2773)
Except for the removal of `collect_basic_blocks`, this just moves code
around and renames.
