+++ openvm/tests/optimizer.rs
@@ -1,4 +1,4 @@
-use powdr_autoprecompiles::optimizer::optimize;
+use powdr_autoprecompiles::constraint_optimizer::optimize_constraints;
 use powdr_autoprecompiles::powdr::UniqueColumns;
 use powdr_autoprecompiles::SymbolicMachine;
 use powdr_number::BabyBearField;
@@ -31,7 +31,7 @@ fn test_optimize() {
     let reader = std::io::BufReader::new(file);
     let machine: SymbolicMachine<BabyBearField> = serde_cbor::from_reader(reader).unwrap();
 
-    let machine = optimize(
+    let machine = optimize_constraints(
         machine,
         OpenVmBusInteractionHandler::new(BusMap::openvm_base()),
         5,
